Windows: fix build, avoiding C# default params (bgo#699379)
When using the latest versions of Mono and .NET (i.e.: the ones
that include the .NET 4.5 profile), even if you decide to use the
.NET 3.5 Profile (whose C# language version was 3.0), it allows
you use language features from higher versions of the language.

This is the reason why the build wasn't failing on mono even when
using the 'gmcs' compiler. However, we still advertise in our
dependencies only the .NET 2.0 Profile, so a Windows developer
could try to build Banshee with a version of Visual Studio which
still doesn't support the "optional parameters" language feature.

We fix this by providing simply additional overloads that account
for the cases in which the parameter is not supplied, in a way in
which we don't break any API.

To avoid this problem arising in the future, we also add the proper
flag to pass to the Mono compiler to error on language features
that are too new for the .NET profile that we target. However, in
the very near future we should probably just switch to .NET 4.0
to make things more simple (not in the stable branch though, which
is where this commit is going to be cherrypicked too).

Signed-off-by: Andrés G. Aragoneses <knocte@gmail.com>